+++ src/hooks/useFetch/index.ts
@@ -1,3 +1,3 @@
-import useFetch from './useFetch';
+import useFetch from './use-fetch';
 
 export default useFetch;
